--- brightness_augmentation3_2.py
@@ -1,508 +1,0 @@
-# -*- coding: utf-8 -*-
-"""brightness_augmentation.ipynb
-
-Automatically generated by Colaboratory.
-
-Original file is located at
-    https://colab.research.google.com/drive/1Q73PMvf8CWGDPXI0b_AJBHL_ksT3UuL5
-"""
-
-!pip install tensorflow_addons
-import numpy as np
-import tensorflow as tf
-from tensorflow import keras
-from tensorflow.keras import layers
-import tensorflow_addons as tfa
-
-
-import pandas as pd
-import json
-import zipfile
-import os
-import seaborn as sns
-import random
-import shutil
-import time
-
-from PIL import Image
-from matplotlib import pyplot as plt 
-
-from keras.models import Sequential, Model
-
-from tensorflow.keras.applications import InceptionV3, Xception, InceptionResNetV2, ResNet50
-from tensorflow.keras.applications.resnet import preprocess_input
-
-from tensorflow.keras.layers import Dense, Conv2D, Flatten, MaxPooling2D, Dropout, GlobalAveragePooling2D
-from tensorflow.keras.preprocessing.image import ImageDataGenerator
-from tensorflow.keras.callbacks import ModelCheckpoint, TensorBoard
-
-!pip install wandb
-import wandb
-from wandb.keras import WandbCallback
-
-!mkdir output
-!mkdir output/tmp-augmented-images/
-
-random.seed(123)
-import tensorflow as tf
-from tensorflow import keras
-from keras import utils
-from keras.preprocessing.image import ImageDataGenerator
-from tensorflow.keras.utils import img_to_array,load_img,array_to_img
-from pathlib import Path
-
-import json
-import zipfile
-import os
-
-kaggle = {
-    "username": "sajacky",
-    "api_key": "5a36788e6cc079b7e25c0712f9f15beb",
-    "on_kernel": False,
-    "dataset": {
-        "sample": "-",
-        "full": "paultimothymooney/chest-xray-pneumonia"
-        }
-}
-
-if kaggle["on_kernel"]:
-  path_prefix = '/kaggle/working/'
-else:
-  path_prefix = '/content/'
-
-# Download dataset
-def download_dataset(which_dataset):
-  data = {"username": kaggle["username"],"key": kaggle["api_key"]}
-  with open('kaggle.json', 'w') as json_file:
-      json.dump(data, json_file)
-
-  !mkdir -p ~/.kaggle
-  !cp kaggle.json ~/.kaggle/
-  !chmod 600 ~/.kaggle/kaggle.json
-  kaggle_dataset = kaggle["dataset"][which_dataset]
-  !kaggle datasets download -d $kaggle_dataset
-  
-  # Paths neeeds to be changed manually because of different directory structures, check with !ls
-  if not os.path.isdir('dataset'):
-    print("Unzipping... ")
-    zip_ref = zipfile.ZipFile('chest-xray-pneumonia.zip', 'r')
-    zip_ref.extractall('dataset')
-    zip_ref.close()
-    !rm -rf /content/dataset/chest_xray/chest_xray
-    !rm -rf /content/dataset/chest_xray/__MACOSX
-    !rm -rf /content/chest-xray-pneumonia.zip
-  
-
-  #!ls Data
-
-
-download_dataset("full")
-
-data_files = os.listdir("dataset/chest_xray")
-print(data_files)
-
-pneumonia_folder ='/content/dataset/chest_xray/train/PNEUMONIA/'
-path,dirs,files = next(os.walk(pneumonia_folder))
-noOfPneumonia = len(files)
-
-normal_folder ='/content/dataset/chest_xray/train/NORMAL/'
-path,dirs,files = next(os.walk(normal_folder))
-noOfNormal = len(files)
-print(noOfPneumonia, noOfNormal)
-
-import tensorflow as tf
-from tensorflow import keras
-from keras import utils
-from keras.preprocessing.image import ImageDataGenerator
-from tensorflow.keras.utils import img_to_array,load_img,array_to_img
-from pathlib import Path
-
-path,dirs,files = next(os.walk('/content/dataset/chest_xray/train/NORMAL/'))
-file_count = len(files)
-print(file_count)
-
-if not os.path.exists('/content/augmented2/'):
-    os.makedirs('/content/augmented2/')
-
-datagen = ImageDataGenerator(
-    brightness_range=[0.5,1.5],
-)
-original_folder ='/content/dataset/chest_xray/train/NORMAL/'
-
-#y=122
-for y in range(0,1242):
-  filename = os.listdir(original_folder)[y]
-  img_path = original_folder+filename
-
-  img = load_img(img_path)
-  x = img_to_array(img)
-  x = x.reshape((1,) + x.shape)
-
-  for batch in datagen.flow(x, batch_size=1, save_to_dir='/content/augmented2/', save_prefix='NORMAL_A', save_format='jpeg'):
-    break
-  
-  y+=1
-  print(y)
-
-path,dirs,files = next(os.walk('/content/augmented2/'))
-file_count = len(files)
-print(file_count)
-
-import shutil
-
-images = [f for f in os.listdir('/content/augmented2/')]
-
-
-for image in images:
-    new_path = '/content/dataset/chest_xray/train/NORMAL/'+image
-    print(new_path)
-    shutil.move('/content/augmented2/'+image, new_path)
-
-CLASSES = os.listdir('/content/dataset/chest_xray/train') 
-TRAINING_DATA_SET_PATH = '/content/dataset/chest_xray/train'
-TEST_DATA_SET_PATH = '/content/dataset/chest_xray/test'
-
-params = dict(
-    seed = 123,
-    image_dim = (288,288),
-    weight_decay = 1e-4,
-    epochs = 30,
-    batch_size = 16,
-    patch_size = 18,
-    pool_size = (2,2),
-    optimizer = 'Adam',
-    l_rate = 0.001,
-    val_split = .15,
-    use_transfer_learning = False,
-    use_data_aug = False,
-
-    l2_reg = .0,
-    projection_dim = 16,
-    num_heads = 4,
-    
-    # Size of the transformer layers
-    transformer_layers = 4,
-    num_classes = len(CLASSES),
-    mlp_head_units = [1024,512]
-    
-    )
-
-new_params = dict(
-    num_patches = (params['image_dim'][0] // params['patch_size']) ** 2,
-    transformer_units = [
-    params['projection_dim'] * 2,
-    params['projection_dim']],
-    input_shape = (params['image_dim'][0], params['image_dim'][1], 3),
-
-)
-params.update(new_params)
-
-
-if params['use_data_aug']:
-  data_aug_params = dict(
-      da_rotation = 20,
-      da_w_shift = 0.1,
-      da_h_shift = 0.1,
-      da_shear = 0.05,
-      da_zoom = 0.05,
-      da_h_flip = False,
-      da_v_flip = False,
-  )
-
-  params.update(data_aug_params)
-
-
-# Ability to switch amount of channel to utilise pre-trained models with specific input shapes
-if params['use_transfer_learning']:
-  INPUT_SHAPE = (params['image_dim'][0], params['image_dim'][1], 3)
-  COLOUR_MODE = 'rgb'
-else:
-  INPUT_SHAPE = (params['image_dim'][0], params['image_dim'][1], 3)
-  COLOUR_MODE = 'rgb'
-
-if params['use_data_aug']:
-  datagen = ImageDataGenerator(validation_split=params['val_split'], rescale=1./255,  
-                               rotation_range=params['da_rotation'],
-                               width_shift_range=params['da_w_shift'],
-                               height_shift_range=params['da_h_shift'],
-                               shear_range=params['da_shear'],
-                               zoom_range=params['da_zoom'],
-                               horizontal_flip=params['da_h_flip'],
-                               vertical_flip=params['da_v_flip'],
-                               fill_mode="constant",
-                               cval=0
-                               )
-else:
-  datagen = ImageDataGenerator(validation_split=params['val_split'])
-
-# Read all training and validation data into variables from directory. 
-# Due to faulty quality of the given validation-set images, all images are taken from the training folder
-train_generator = datagen.flow_from_directory(TRAINING_DATA_SET_PATH,
-                                                    batch_size=params['batch_size'],
-                                                    seed=123,
-                                                    class_mode="categorical",
-                                                    classes=CLASSES,
-                                                    target_size=params['image_dim'],
-                                                    color_mode=COLOUR_MODE,
-                                                    subset='training',
-                                                    shuffle=True)
-
-val_datagen = ImageDataGenerator(validation_split=0.15, rescale=1./255)
-valid_generator = datagen.flow_from_directory(TRAINING_DATA_SET_PATH,
-                                                    batch_size=params['batch_size'],
-                                                    seed=123,
-                                                    class_mode="categorical",
-                                                    classes=CLASSES,
-                                                    target_size=params['image_dim'],
-                                                    color_mode=COLOUR_MODE,
-                                                    subset='validation',
-                                                    shuffle=False)
-
-test_datagen = ImageDataGenerator(rescale=1./255)
-test_generator = test_datagen.flow_from_directory(TEST_DATA_SET_PATH,
-                                                    batch_size=params['batch_size'],
-                                                    seed=123,
-                                                    class_mode="categorical",
-                                                    classes=CLASSES,
-                                                    target_size=params['image_dim'],
-                                                    color_mode=COLOUR_MODE,
-                                                    shuffle=False)
-
-def mlp(x, hidden_units, dropout_rate):
-    for units in hidden_units:
-        x = layers.Dense(units, activation=tf.nn.gelu)(x)
-        x = layers.Dropout(dropout_rate)(x)
-    return x
-
-class Patches(layers.Layer):
-    def __init__(self, patch_size):
-        super(Patches, self).__init__()
-        self.patch_size = patch_size
-
-    def call(self, images):
-        batch_size = tf.shape(images)[0]
-        patches = tf.image.extract_patches(
-            images=images,
-            sizes=[1, self.patch_size, self.patch_size, 1],
-            strides=[1, self.patch_size, self.patch_size, 1],
-            rates=[1, 1, 1, 1],
-            padding="VALID",
-        )
-        patch_dims = patches.shape[-1]
-        patches = tf.reshape(patches, [batch_size, -1, patch_dims])
-        return patches
-
-    
-# Linearly transform patches by projecting it into a
-# vector of size `projection_dim` and also adds a learnable position
-# embedding to the projected vector.
-class PatchEncoder(layers.Layer):
-    def __init__(self, num_patches, projection_dim):
-        super(PatchEncoder, self).__init__()
-        self.num_patches = num_patches
-        self.projection = layers.Dense(units=projection_dim)
-        self.position_embedding = layers.Embedding(
-            input_dim=num_patches, output_dim=projection_dim
-        )
-
-    def call(self, patch):
-        positions = tf.range(start=0, limit=self.num_patches, delta=1)
-        encoded = self.projection(patch) + self.position_embedding(positions)
-        return encoded
-
-
-import matplotlib.pyplot as plt
-
-plt.figure(figsize=(4, 4))
-image, label = iter(next(train_generator))
-image = image[0]*255.
-#image = x_train[np.random.choice(range(x_train.shape[0]))]
-plt.imshow((image).astype("uint8"))
-plt.axis("off")
-
-resized_image = tf.image.resize(
-    tf.convert_to_tensor([image]), size=(params['image_dim'][0], params['image_dim'][0])
-)
-patches = Patches(params['patch_size'])(resized_image)
-print(f"Image size: {params['image_dim'][0]} X {params['image_dim'][0]}")
-print(f"Patch size: {params['patch_size']} X {params['patch_size']}")
-print(f"Patches per image: {patches.shape[1]}")
-print(f"Elements per patch: {patches.shape[-1]}")
-
-n = int(np.sqrt(patches.shape[1]))
-plt.figure(figsize=(4, 4))
-for i, patch in enumerate(patches[0]):
-    ax = plt.subplot(n, n, i + 1)
-    patch_img = tf.reshape(patch, (params['patch_size'], params['patch_size'], 3))
-    plt.imshow(patch_img.numpy().astype("uint8"))
-    plt.axis("off")
-
-def create_vit_classifier():
-    inputs = layers.Input(shape=params['input_shape'])
-    # Create patches.
-    patches = Patches(params['patch_size'])(inputs)
-    # Encode patches.
-    encoded_patches = PatchEncoder(params['num_patches'], params['projection_dim'])(patches)
-
-    # Create multiple layers of the Transformer block.
-    for _ in range(params['transformer_layers']):
-        # Layer normalization 1.
-        x1 = layers.LayerNormalization(epsilon=1e-6)(encoded_patches)
-        # Create a multi-head attention layer.
-        attention_output = layers.MultiHeadAttention(
-            num_heads=params['num_heads'], key_dim=params['projection_dim'], dropout=0.1
-        )(x1, x1)
-        # Skip connection 1.
-        x2 = layers.Add()([attention_output, encoded_patches])
-        # Layer normalization 2.
-        x3 = layers.LayerNormalization(epsilon=1e-6)(x2)
-        # MLP.
-        x3 = mlp(x3, hidden_units=params['transformer_units'], dropout_rate=0.1)
-        # Skip connection 2.
-        encoded_patches = layers.Add()([x3, x2])
-
-    # Create a [batch_size, projection_dim] tensor.
-    representation = layers.LayerNormalization(epsilon=1e-6)(encoded_patches)
-    representation = layers.Flatten()(representation)
-    representation = layers.Dropout(0.5)(representation)
-    # Add MLP.
-    features = mlp(representation, hidden_units=params['mlp_head_units'], dropout_rate=0.5)
-    # Classify outputs.
-    logits = layers.Dense(len(CLASSES))(features)
-    # Create the Keras model.
-    model = keras.Model(inputs=inputs, outputs=logits)
-    return model
-
-from keras import backend as K
-
-# For experiment tracking
-USE_WANDB = False
-
-if USE_WANDB:
-  wandb.init(tags=["ViT", "Binary"], project='Ashik babus project', entity='Ashik', sync_tensorboard=True, config=params, save_code=True)
-  print(params)
-  wandb_callback = WandbCallback(monitor='val_f1_m', 
-                                 save_model=True,
-                                 save_weights_only=False, mode='max',
-                                 log_weights=True,
-                                 data_type="image", verbose=1, 
-                                 labels=CLASSES, 
-                                 generator=valid_generator,
-                                 predictions=50,
-                                 log_evaluation=True,
-                                 log_batch_frequency=1) 
-
-def run_experiment(model):
-    optimizer = tfa.optimizers.AdamW(
-        learning_rate=params['l_rate'], weight_decay=params['weight_decay']
-    )
-    
-    def recall_m(y_true, y_pred):
-        true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
-        possible_positives = K.sum(K.round(K.clip(y_true, 0, 1)))
-        recall = true_positives / (possible_positives + K.epsilon())
-        return recall
-
-    def precision_m(y_true, y_pred):
-        true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
-        predicted_positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
-        precision = true_positives / (predicted_positives + K.epsilon())
-        return precision
-    
-    def f1_m(y_true, y_pred):
-        precision = precision_m(y_true, y_pred)
-        recall = recall_m(y_true, y_pred)
-        return 2*((precision*recall)/(precision+recall+K.epsilon()))
-
-    model.compile(
-        optimizer=optimizer,
-        loss=keras.losses.BinaryCrossentropy(from_logits=True),
-        metrics=[
-            keras.metrics.BinaryAccuracy(name="binary_accuracy"), f1_m, recall_m, precision_m
-        ],
-    )
-
-    
-    checkpoint_callback = keras.callbacks.ModelCheckpoint(
-        "/content/output/model_best.h5",
-        monitor="val_f1_m",
-        save_best_only=True,
-        save_weights_only=True,
-        verbose=1,
-        mode="max"
-    )
-
-    callbacks_list = [checkpoint_callback, tf.keras.callbacks.EarlyStopping(patience=20, monitor="val_binary_accuracy")]
-
-    if USE_WANDB:
-      callbacks_list.append(wandb_callback)
-
-
-    history = model.fit(
-        train_generator,
-        batch_size=params['batch_size'],
-        epochs=params['epochs'],
-        validation_data=valid_generator,
-        callbacks=callbacks_list) 
-
-    return history, model
-
-model = create_vit_classifier()
-history, model = run_experiment(model)
-
-import matplotlib.pyplot as plt
-
-def plot_precision_and_accuracy(history):
-    acc = history.history['binary_accuracy']
-    val_acc = history.history['val_binary_accuracy']
-    precision = history.history['precision_m']
-    val_precision = history.history['val_precision_m']
-    epochs_range = range(len(history.epoch))
-    
-    plt.figure(figsize=(15, 5))
-    plt.subplot(1, 2, 1)
-    plt.plot(epochs_range, acc, label='Training Accuracy')
-    plt.plot(epochs_range, val_acc, label='Validation Accuracy')
-    plt.legend(loc='lower right')
-    plt.title('Training and Validation Accuracy')
-
-    plt.subplot(1, 2, 2)
-    plt.plot(epochs_range, precision, label='Training Precision')
-    plt.plot(epochs_range, val_precision, label='Validation Precision')
-    plt.legend(loc='lower right')
-    plt.title('Training and Validation Precision')
-
-    plt.show()
-    
-plot_precision_and_accuracy(history)
-
-import csv
-import datetime
-import os
-
-# Get the metric values from the history object
-acc = history.history['binary_accuracy']
-val_acc = history.history['val_binary_accuracy']
-precision = history.history['precision_m']
-val_precision = history.history['val_precision_m']
-
-# Specify the directory for the CSV file
-directory = './results'
-
-# Create the directory if it does not exist
-if not os.path.exists(directory):
-    os.makedirs(directory)
-
-# Create a new CSV file with timestamp and write the metric values to it
-filename = os.path.join(directory, 'brightness_3_2' + datetime.datetime.now().strftime("%Y%m%d_%H%M%S") + '.csv') #change 'orignal' to method of augmentation tried.
-with open(filename, mode='w') as metrics_file:
-    fieldnames = ['epoch', 'accuracy', 'val_accuracy', 'precision', 'val_precision']
-    writer = csv.DictWriter(metrics_file, fieldnames=fieldnames)
-
-    writer.writeheader()
-    for epoch in range(len(acc)):
-        writer.writerow({'epoch': epoch+1, 'accuracy': acc[epoch], 'val_accuracy': val_acc[epoch], 'precision': precision[epoch], 'val_precision': val_precision[epoch]})
-
-
-dir_to_save = '/content/drive/MyDrive/Main Project/'
-shutil.copyfile(filename,dir_to_save+'brightness32.csv')
